pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medianav-toolbox"
version = "0.1.0"
description = "Linux/Python replacement for the Dacia MediaNav Evolution Toolbox"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "httpx>=0.27",
    "click>=8.1",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "respx>=0.21",
    "black>=24.0",
    "isort>=5.13",
]

[project.scripts]
medianav-toolbox = "medianav_toolbox.cli:cli"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: marks tests that hit the real API (deselect with '-m not integration')"]

[tool.setuptools.packages.find]
include = ["medianav_toolbox*"]
